IsoMorphic.java
We first create a one to one mapping (map 2) of characters from string t to string s so that we can keep record of which character in string t will replace which character of string s.

then we iterate through string s . We check if that character has already been replaced by character of string t . 

Time - O(n)
Space- O(n)
If the character has not been replaced i.e it is not in map, then we create a new entry for this , check whether replaced character has already replaced some other character using map2 and add replaced character to string builder.
if the character has already been replaced i.e it is already in map , then we simply replace it with that character of string t which was previously used to replace the same character earlier in string s. The character gets added to string builder.

In the end we check if string from stringbuilder is equal to string t or not.