Problem 1 - Anagrams
We take each string, convert it into a character array and sort it so that all anagrams become the same string.
We add this as a key to hashmap and append an arraylist as a value to this key. The string gets appended to the arraylist.
if this key is already present , simply append the string to the arraylist at that key.

In the end we return an array list of all the values from the map.